+++ voom-service/src/main/java/inc/visor/voom_service/VoomServiceApplication.java
@@ -2,6 +2,7 @@ package inc.visor.voom_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
