+++ voom-service/src/main/java/inc/visor/voom_service/auth/user/repository/UserRoleRepository.java
@@ -1,8 +1,9 @@
 package inc.visor.voom_service.auth.user.repository;
 
+import inc.visor.voom_service.auth.user.model.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRoleRepository extends JpaRepository<UserRoleRepository, Integer> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
 }
