+++ voom-service/src/main/java/inc/visor/voom_service/config/WebSocketConfig.java
@@ -11,9 +11,7 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-
         registry.addEndpoint("/ws").withSockJS();
-
     }
 
     @Override
