+++ voom-service/src/main/java/inc/visor/voom_service/vehicle/controller/VehicleTypeController.java
@@ -28,7 +28,7 @@ public class VehicleTypeController {
         return ResponseEntity.ok(new VehicleType());
     }
 
-    @PostMapping("/{id}")
+    @PutMapping("/{id}")
     public ResponseEntity<VehicleType> updateVehicleType(@PathVariable long id, @RequestBody VehicleType vehicleTypeDto) {
         return ResponseEntity.ok(vehicleTypeDto);
     }
