+++ voom-service/src/main/resources/application.properties
@@ -1,14 +1,12 @@
 spring.application.name=voom-service
 
-spring.autoconfigure.exclude=\
-org.springframework.boot.autoconfigure.data.jpa.JpaRepositoriesAutoConfiguration,\
-org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration
-
 # Database configuration
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 spring.datasource.username=postgres
 spring.datasource.password=secret
-spring.jpa.hibernate.ddl-auto=update
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 jwt.secret=xrMUZBaXAbciQVziPdT1zrGzT9jr16OoASeFFQ+b/BQt/0A60yCEAHoxUucJt6VQ
